backend/chat-gpt-api/requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
pydantic